+++ utilities/source_utilities.py
@@ -211,7 +211,10 @@ def s3_publish_image(id, client):
         client.upload_file(
             '{}/{}.jpg'.format(current_app.config['WEBTOOL_IMAGE_MOUNT_POINT'], id),
             current_app.config['BUCKET'],
-            '{}/{}/{}.jpg'.format(current_app.config['AWS_IMAGE_FOLDER_PREFIX'], id, imageId)
+            '{}/{}/{}.jpg'.format(current_app.config['AWS_IMAGE_FOLDER_PREFIX'], id, imageId),
+            ExtraArgs={
+                'ACL': 'bucket-owner-full-control'
+            }
         )
 
     except Exception: